Check for storage type before clear database

+++ public_html/js/app/models/User.js
@@ -81,7 +81,8 @@ define([
                     self.set(result);
                     self.settings.set('id', result.user_id);
                     self.settings.fetch(function() {
-                        window.indexedDB.deleteDatabase(result.user_id);
+                        if (skritter.settings.isIndexedDB())
+                            window.indexedDB.deleteDatabase(result.user_id);
                         localStorage.setItem('active', result.user_id);
                         callback(result);
                     });
